requirements.txt
numpy~=2.0.2
matplotlib~=3.9.4
opencv-python~=4.12.0.88
pandas~=2.3.1
torch~=2.7.1
torchvision~=0.22.1
pillow~=11.2.1
torchviz~=0.0.3
xlwt~=1.3.0
tqdm~=4.67.1
timm~=1.0.15